Update attributes for managing user roles
Replace all attributes that were previously named "type" or "typename" with new names "role" and "roleName" respectively. This change provides better description for user roles and are less likely to conflict with Go's built-in "type" keyword. Also, it now includes the user role information in the JWT claims when a token is issued. Adjusted in various repositories, handlers and attributes for compatibility. The change in attribute name also reflects in the MongoDB queries. Updated swagger documentation accordingly.